requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
passlib[bcrypt]
pydantic[email]
PyJWT